added more parameters in location and computed properties for icon's url
